+++ tensorcircuit/interfaces.py
@@ -133,24 +133,44 @@ def torch_interface(fun: Callable[..., Any], jit: bool = False) -> Callable[...,
 
 
 def scipy_optimize_interface(
-    fun: Callable[..., Any], shape: Optional[Tuple[int, ...]] = None, jit: bool = True
+    fun: Callable[..., Any],
+    shape: Optional[Tuple[int, ...]] = None,
+    jit: bool = True,
+    gradient: bool = True,
 ) -> Callable[..., Any]:
-    vag = backend.value_and_grad(fun, argnums=0)
-    if jit:
-        vag = backend.jit(vag)
+    if gradient:
+        vag = backend.value_and_grad(fun, argnums=0)
+        if jit:
+            vag = backend.jit(vag)
 
-    def scipy_vag(*args: Any, **kws: Any) -> Tuple[Tensor, Tensor]:
+        def scipy_vag(*args: Any, **kws: Any) -> Tuple[Tensor, Tensor]:
+            scipy_args = numpy_args_to_backend(args, dtype=dtypestr)
+            if shape is not None:
+                scipy_args = list(scipy_args)
+                scipy_args[0] = backend.reshape(scipy_args[0], shape)
+                scipy_args = tuple(scipy_args)
+            vs, gs = vag(*scipy_args, **kws)
+            scipy_vs = general_args_to_numpy(vs)
+            gs = backend.reshape(gs, [-1])
+            scipy_gs = general_args_to_numpy(gs)
+            scipy_vs = scipy_vs.astype(np.float64)
+            scipy_gs = scipy_gs.astype(np.float64)
+            return scipy_vs, scipy_gs
+
+        return scipy_vag
+    # no gradient
+    if jit:
+        fun = backend.jit(fun)
+
+    def scipy_v(*args: Any, **kws: Any) -> Tensor:
         scipy_args = numpy_args_to_backend(args, dtype=dtypestr)
         if shape is not None:
             scipy_args = list(scipy_args)
             scipy_args[0] = backend.reshape(scipy_args[0], shape)
             scipy_args = tuple(scipy_args)
-        vs, gs = vag(*scipy_args, **kws)
+        vs = fun(*scipy_args, **kws)
         scipy_vs = general_args_to_numpy(vs)
-        gs = backend.reshape(gs, [-1])
-        scipy_gs = general_args_to_numpy(gs)
         scipy_vs = scipy_vs.astype(np.float64)
-        scipy_gs = scipy_gs.astype(np.float64)
-        return scipy_vs, scipy_gs
+        return scipy_vs
 
-    return scipy_vag
+    return scipy_v